+++ src/account/auth/new_wallet.rs
