+++ src/account/auth/test_notifs.rs
@@ -1,5 +1,6 @@
 use crate::{
-    constants::merkle::API_ROOT, types::auth::secret::Secret, types::v2::user::UserRoot, Farcaster,
+    constants::merkle::API_ROOT, types::account::auth::secret::Secret, types::v2::user::UserRoot,
+    Farcaster,
 };
 
 impl Farcaster {
